renaming fiscal year in readme and spec
