+++ README.md
@@ -104,11 +104,11 @@ retail_calendar.end_of_week(2017, 4, 1)
 
 ### Offset fiscal year calendars
 Some companies, one of which being Stitch Fix, operate on a fiscal year calendar that is offset from
-the traditional retail calendar.  The `MerchCalendar::FiscalYearCalendar` class allows you to easily
+the traditional retail calendar.  The `MerchCalendar::StitchFixFiscalYearCalendar` class allows you to easily
 offset the start of year to match your fiscal calendar.
 
 ```ruby
-fiscal_calendar = MerchCalendar::FiscalYearCalendar.new
+fiscal_calendar = MerchCalendar::StitchFixFiscalYearCalendar.new
 
 # 52 or 53 (depending on leap year)
 fiscal_calendar.weeks_in_year(2017)
