+++ lib/merch_calendar/stitch_fix_fiscal_year_calendar.rb
@@ -1,5 +1,6 @@
 module MerchCalendar
   class StitchFixFiscalYearCalendar
+
     QUARTER_1 = 1
     QUARTER_2 = 2
     QUARTER_3 = 3
