+++ lib/merch_calendar/util.rb
@@ -1,5 +1,5 @@
 require "merch_calendar/retail_calendar"
-require "merch_calendar/fiscal_year_calendar"
+require "merch_calendar/stitch_fix_fiscal_year_calendar"
 
 module MerchCalendar
 
@@ -179,7 +179,7 @@ module MerchCalendar
     end
 
     def fiscal_year_calendar
-      @fiscal_year_calendar ||= FiscalYearCalendar.new
+      @fiscal_year_calendar ||= StitchFixFiscalYearCalendar.new
     end
 
     # Reads the provided parameter and converts the value
